feat(stats): faceted advanced charts + extended range selector
Implements #6 and #7.

- Stats: range selector extended to day/week/month/3m/6m/1y/all,
  defaulting to the last 3 months so the page only fetches a bounded
  window (#7); last selected range remembered per session
- Stats: opt-in advanced mode replaces the static grid with a faceted
  dashboard — live advanced-search filters (keyword, favorite, time,
  mood, recall, type, symbols) drive a full-width timeline with
  range-based buckets (daily/weekly/monthly) plus time/type/symbols/
  recall/mood mini-charts computed from the filtered dreams, and a
  "View as list" deep-link into Search (#6); filters persist per
  session; static charts unchanged when opted out
- Settings: new "Advanced charting" card toggling the opt-in
- db: additive User.advancedCharting column (migration
  20260702155651_advanced_charting_opt_in; deploy with
  `blitz prisma migrate deploy`)
- refactor: search where-clause extracted into shared
  buildDreamSearchWhere (reused by Search and advanced stats);
  setChartsData moved to src/stats/helpers (reused by the facet
  charts and the PDF export)
- docs: CLAUDE.md context files (root/db/src) incl. frozen-deps
  policy, plus PLAN-issues-6-7.md

No new dependencies.
