+++ src/settings/components/ExportDreams/index.tsx
@@ -6,7 +6,7 @@ import { Fragment, useMemo, useState } from "react"
 import getDreams from "src/dreams/queries/getDreams"
 import getSymbolsWithUsage from "src/symbols/queries/getSymbolsWithUsage"
 import { StatGoogleChart } from "src/stats/components/StatGoogleChart"
-import { setChartsData } from "src/pages/stats"
+import { setChartsData } from "src/stats/helpers/chartsData"
 import { StatSymbolChart } from "src/stats/components/StatSymbolChart"
 import LoadingSpiral from "src/core/components/LoadingSpiral"
 import { useCurrentUser } from "src/core/hooks/useCurrentUser"
