+++ src/stats/helpers/range.ts
@@ -1,5 +1,9 @@
 export type Range = "day" | "week" | "month" | "3months" | "6months" | "1year" | "all"
 
+// sessionStorage keys: remembered per browser tab/session, clean slate on a fresh visit
+export const STATS_RANGE_STORAGE_KEY = "dreamingsheep.stats.range"
+export const ADVANCED_STATS_FILTERS_STORAGE_KEY = "dreamingsheep.stats.advancedFilters"
+
 // days to subtract from "today"; null = no lower bound (all time)
 export const RANGE_TO_DAYS: Record<Range, number | null> = {
   day: 0,
